Fix #5 Add Bootlint
If you already have a `.lints.json`, in order to use Bootlint you
need to add it to your `.lints.json`, as in
[src/default.lints.json](src/default.lints.json).

If you wish to disable some of the Bootlint's linters you can
create a `.bootlintrc.json` file in the root of your project with
an array of string linter IDs to disable, e. g.

```json
["W001", "W002", "W003"]
```